pyproject.toml
[project]
name = "e-commerce-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "duckduckgo-search>=8.0.4",
    "geopandas>=1.1.0",
    "google>=3.0.0",
    "google-genai>=1.20.0",
    "google-generativeai>=0.8.5",
    "kaleido>=1.0.0",
    "langchain>=0.3.26",
    "langchain-core>=0.3.66",
    "langchain-google-genai>=2.0.10",
    "langfuse>=3.0.3",
    "langgraph>=0.4.8",
    "litellm>=1.72.6.post1",
    "llama-index-embeddings-google-genai>=0.2.0",
    "llama-index-embeddings-huggingface>=0.5.4",
    "llama-index-llms-google-genai>=0.2.1",
    "llama-index-llms-huggingface-api>=0.5.0",
    "llama-index-tools-google>=0.3.1",
    "llama-index-utils-workflow>=0.3.5",
    "llama-index-vector-stores-chroma>=0.4.2",
    "mcp>=1.9.4",
    "openinference-instrumentation-smolagents>=0.1.13",
    "opentelemetry-exporter-otlp>=1.34.1",
    "opentelemetry-sdk>=1.34.1",
    "pillow>=11.2.1",
    "plotly>=6.1.2",
    "python-dotenv>=1.1.0",
    "requests>=2.32.4",
    "shapely>=2.1.1",
    "smolagents[litellm]>=1.18.0",
]
